+++ server.go
@@ -464,6 +464,7 @@ func (s *Server) StartSelfMonitoring(database, retention string, interval time.D
 
 			// Shard-level stats.
 			tags["shardID"] = strconv.FormatUint(s.id, 10)
+			s.mu.RLock()
 			for _, sh := range s.shards {
 				if !sh.HasDataNodeID(s.id) {
 					// No stats for non-local shards.
@@ -471,6 +472,7 @@ func (s *Server) StartSelfMonitoring(database, retention string, interval time.D
 				}
 				batch = append(batch, pointsFromStats(sh.stats, tags)...)
 			}
+			s.mu.RUnlock()
 
 			// Server diagnostics.
 			for _, row := range s.DiagnosticsAsRows() {
@@ -2141,9 +2143,20 @@ func (s *Server) createShardGroupsIfNotExists(database, retentionPolicy string, 
 		// Local function makes locking fool-proof.
 		s.mu.RLock()
 		defer s.mu.RUnlock()
+
+		rp, err := s.RetentionPolicy(database, retentionPolicy)
+		if err != nil {
+			return err
+		}
+
+		times := map[time.Time]struct{}{}
 		for _, p := range points {
+			times[p.Time.Truncate(rp.ShardGroupDuration)] = struct{}{}
+		}
+
+		for t := range times {
 			// Check if shard group exists first.
-			g, err := s.shardGroupByTimestamp(database, retentionPolicy, p.Time)
+			g, err := s.shardGroupByTimestamp(database, retentionPolicy, t)
 			if err != nil {
 				return err
 			} else if g != nil {
@@ -2152,7 +2165,7 @@ func (s *Server) createShardGroupsIfNotExists(database, retentionPolicy string, 
 			commands = append(commands, &createShardGroupIfNotExistsCommand{
 				Database: database,
 				Policy:   retentionPolicy,
-				Time:     p.Time,
+				Time:     t,
 			})
 		}
 		return nil
@@ -2426,6 +2439,8 @@ func (s *Server) rewriteSelectStatement(stmt *influxql.SelectStatement) (*influx
 			return nil, err
 		}
 	}
+
+	stmt.RewriteDistinct()
 
 	return stmt, nil
 }